Add shortcut to spawn new terminal in the current dir
Ctrl-Shift-Return now creates a new ST terminal, whose CWD is the same
as the parent st's CWD.

This version of the patch does a double fork, a technique commonly used
by daemons to spawn orphan processes.

This solution is specific to the swallow patch for dwm which traverses
the process tree to determine if the new window is a decendant of a
terminal window, in which case the new window should take the place of
the terminal window.

The way the original newterm patch worked the new st terminal would be
a direct decendant of the parent st terminal process, which could lead
to the wrong terminal window being swallowed.

The double fork method avoids this by leaving all new st terminals as
orphans, i.e. they will have no parent process.